myproject/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==2.0.16
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.5.0
APScheduler==3.10.4
